Update resolveInvokeDynamic to use ramClass
In cases where class redefinition causes an old version of a method to
be active on the stack, using the class Object to resolve invokeDynamic
calls will result in searching cpIndexes with the wrong constantPool
since the classObject always points to the newest version
of the class.

This change modifies 'resolveInvokeDynamic' so that it uses the ramClass
instead of the classObject so it always looks up indexes in the correct
constantPool.

Signed-off-by: tajila <atobia@ca.ibm.com>