Add support for 64-bit synchronized field load
Enhance ReduceSynchronizedFieldLoad with 64-bit field loading support.
This feature further improves the optimization with additional asserts
around the use of the LPDG instruction. All 8 possibilities are
considered:

1. 31-bit JVM loading a 32-bit field
2. 31-bit JVM loading a 64-bit field
3. 64-bit JVM loading a 32-bit field
4. 64-bit JVM loading a 64-bit field
5. 64-bit JVM loading a 32-bit field with a compressed lockword
6. 64-bit JVM loading a 64-bit field with a compressed lockword

Only 2. is not supported due to register pairs. The optimization will
ensure that alignment restrictions of LPD(G) are enforced in all
scenarios.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>