Store offset and scale per type in VH Ops classes instead of VarHandle
Offset and scale are used calculate the memory address given an index.
These are constant, and don't need to be stored in VH instance fields.
Storing in static final fields in ArrayVarHandleOperations.Op* instead.

This should benefit the VM (invokestatic instead of invokevirtual).
It should also benefit the JIT (static final values).

Signed-off-by: Bjørn Vårdal <bjornvar@ca.ibm.com>