search for loadlocationtype when calling defineclass
When calling Class.forname() it uses ClassLoader.defineclass() to load
the class. In this case we never search for where the class was loaded
from and fill the J9TranslationLocalBuffer. This means that when we are
creating the ramClass the loadLocation is set to LOAD_LOCATION_UNKNOWN
and we will not bother to search for the module that the class belongs
to, instead we will assign it to the unamedModule.

This will cause problems as classes which are loaded from modules will
have the incorrect module in the J9Class.

This change ensures that we always search for the class location, if the
class cannot be found we leave the location as LOAD_LOCATION_UNKNOWN.

Also removed the obsolete J2ME flags

Closes: #110

Signed-off-by: tajila <atobia@ca.ibm.com>