Change the last arg for VarHandle operation methods to VarHandles
The argument used to be specific to the VarHandle subtype.
To support adding the VarHandle argument directly at the call site,
the argument must be valid for all VarHandle subtypes.

Other changes partially eliminate the use of the VarHandle arg,
which removes the need for a checkcast.

This is part of the changes necessary to remove the VH typeTable.
This commit also removes the initialization of the typeTable.

Signed-off-by: Bjørn Vårdal <bjornvar@ca.ibm.com>