Add static and dynamic debug counters
Add counters so we can track statistics of this optimization in the
future. We add a single dynamic debug counter in the fallback OOL path
so we have the ability to determine if this optimization is truly
helping by synchronously loading the field without acquiring a lock. It
should be extremely rare to see ticks in the fallback code.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>