Modify xhelpers.m4 and m4 build options to work on OSX
Duplicated M4 UMA rule to be able to pass the `-DOSX` to
m4.  This follows the pattern used by other platforms.

Workaround the OSX assembler not supporting `je short`.
Add a define to the xhelpers.m4 for SHORT_JMP which
is turned into `short` on windows and linux-x86.  On
OSX it defines away to nothing.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>