Improve perf of j/l/Class annotation APIs
The current implementation of the various annoation APIs in
java/lang/Class are based on the AnnotationCache data structure which
has two fields of type Map - an interface type. The maps are either set
to and instance of the EmptyMap type or an instance of LinkedHashMap.
This change switches EmptyMap cases to be null. This in turn allows the
fields to be declared to be of the concrete LinkedHashMap type which
boosts performance and simplifies handling of empty maps. This also
makes the call profiling from the IProfiler much less confusing for the
JIT leading to throughput boost of around 30% in microbenchmarks of the
getAnnotation API.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>