Update VMnewEvaluator to create and use dummy call node for helper call
For 64-bit VM when we pass arguments to JIT helper for new allocations,
we need to sign extend type/size child node of newarray or anewarray node
as C helper linkage uses whole 64-bit register and evaluating 32-bit child
of node only changes lower half of the register. Due to this reason, we
also needed to pass a stack of argument registers to buildDirectDispatch to
generate a helper sequence.
This commit creates and uses a dummy helper call node with appropriate child
to handle above scenario so that we can avoid passing stack of registers around.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>