Allow non-JNI unsafes for Z codegen
In Java9 sun.misc.Unsafe JNI methods are moved to jdk.internal and
replaced with simple wrappers. This leads to a few places in the
code where we do the wrong thing by assuming that the recognized
Unsafe method is a JNI method.
This commit adds explicit tests to prevent z codegen from attempting
to transform the non-JNI versions of the unsafe CAS instructions.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>