Check preferred size on SegmentProvider init
When J9::J9SegmentCache is initialized, it first tries to allocate
a segment of 16 MB; failing that it tries to allocate 2 MB, and
failing that it tries to allocate 64 KB. The problem occurs when it
allocates either 2 MB or 64 KB. When the J9::SystemSegmentProvider is
initialized in TR::CompilationInfoPerThreadBase::compile, it tries to
allocate a 16 MB segment. Because it takes in a J9::J9SegmentCache which
has already preallocated 16 MB, it just receives this preallocated
segment. However, if J9::J9SegmentCache allocated 2 MB or 64 KB, then
when J9::SystemSegmentProvider requests a 16 MB segment, J9::SegmentCache
will see that its preallocated segment is not big enough, and will try to
allocate the requested size, which will most likely fail since it failed
earlier.

This commit adds an additional API to the J9::J9SegmentProvider
interface class in order for consumers to query what the preferred
segment size is. In the case of J9::J9SegmentCache, the preferred
segment size is the size of the cached segment.

When initializing J9::SystemSegmentProvider, it uses the preferred
segment size (if it is non zero) to set its default system segment size
so that when it allocates a segment at initialization time,
J9::J9SegmentCache provides it the one that was already allocated.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>