Allow J9 tools to build on OpenJDK7
The documentation for building OpenJDK8 says to use OpenJDK7 as the
boot java. This means that the tools compilation fails because it
tries to use java 8 as the target level. Java 7 doesn't know about
target 8. Since the tools compile fine on Java 7, we should changing
the SPEC_LEVEL (target) to 7 so we can compile the tools with the same
boot java we use to build OpenJDK8.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>