Fix AbstractMethodError
Java compiler needs to make sure that an abstract method is only
referenced in an abstract class. JVM does not need to enforce this
check during lookup. 

Removing this check should not influence the JVM. Since abstract methods
are real in the vTable, AbstractMethodError will be thrown if an
abstract method is run.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>