Allow the ARM arch flags to be set
In order to be able to compile for either arm6 or arm7 we have to
be able to set the gcc architecture flags for -march=<type> and -marm.
I modified the assignment of ARM_ARCH_FLAGS to use ?= so we can set
the desired value from either the calling makefile or the environment.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>