Prevent use of store multiple for iprofiler data
Store multiple with unaligned data is not supported by ARMv8, so
we have to make sure that gcc does not generate them. It usually
gets this right be default, but the casting we do when packing
iprofiler data confuses the compiler, resulting in a store multiple
that can crash the JVM when running on a 64 bit OS (some but not
all kernels have a trap handler that catches the bus error and
emulates the unaligned access, at very high overhead).

We can prevent gcc from generating the stm instruction by replacing
the assignments with calls to memcpy. These should be inlined, so
the overhead isn't as bad as it sounds.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>