Remove dmb_v6 when running as ARMv7
dmb_v6 is extremely bad when running under a 64 bit OS. Most of them are
replaced by dmb when running on ARMv7, but there is one in VMmonexitEvaluator.
Strangely, the line above is the commented out conditional form used everywhere
else. This change restores that version until we can figure out why it was
not being used.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>