Avoid stm instructions for interpreterProfiling
stm on unaligned data causes sigbus when running armhf under a 64 bit OS,
so we have to force the assignments to be coded separately. This version
uses memcpy to do so, but I would like to get the inline assembly working
for a more transparent and robust solution. (the commented out code causes
crashes)

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>